refactor(cli): extract skills subcommand into cli/commands/skills.mjs
Move the `skills` (resolve | import/export/list | default run) blocks out of
main() into a dedicated handler. Behavior, messages, and exit codes are
unchanged; only the enclosing function and relative import depth differ.
